+++ Sources/Scaletor/Extensions/Chord+CustomStringConvertible.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension Chord: CustomStringConvertible {
+    public var description: String {
+        ChordFormatter().string(from: self)
+    }
+}