+++ Sources/Scaletor/Extensions/StringInterpolation+Chord.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public extension String.StringInterpolation {
+    mutating func appendInterpolation(_ chord: Chord, style: ChordFormatter.Style) {
+        let formatter = ChordFormatter(style: style)
+        appendInterpolation(formatter.string(from: chord))
+    }
+}